Updated For Project Final

+++ backend/src/main/java/com/project/nutriTrack/dto/CalorieTargetRequest.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class CalorieTargetRequest {
-    private int age;
+    private Double age;
     private double weight_kg;
     private double height_cm;
     private String gender;
     private String activity_level;
     private double target_weight_kg;
-    private int time_weeks;
+    private Double time_weeks;
 }
